requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
Automat==20.2.0
babel==2.16.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bitsandbytes==0.44.1
black==24.8.0
bleach==6.1.0
blinker==1.4
blobfile==3.0.0
Brotli==1.0.9
celery==5.4.0
certifi==2020.6.20
cffi==1.17.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-init==24.3.1
colab==1.13.5
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
cryptography==3.4.8
cssselect==1.2.0
cupshelpers==1.0
cycler==0.12.1
datasets==3.1.0
dbus-python==1.2.18
debugpy==1.8.5
decorator==5.1.1
deepspeed==0.15.4
defer==1.0.6
defusedxml==0.7.1
deluge==2.0.3
diazo==2.0.2
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==1.7.11
django-haystack==2.4.1
django-hitcounter==0.1.1
django-revproxy==0.10.0
django-taggit==0.22.2
etiquetando==0.1
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
fire==0.7.0
flake8==7.1.1
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.2.0
GeoIP==1.3.2
greenlet==1.1.2
h11==0.14.0
hjson==3.1.0
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.26.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch==1.32
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
jupytext==1.16.4
keyring==23.5.0
kiwisolver==1.4.5
kombu==5.4.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libtorrent===2.0.5-build-libtorrent-rasterbar-qrM5vM-libtorrent-rasterbar-2.0.5-bindings-python
llama_models==0.0.43
llama_stack==0.0.43
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.1.0
multiprocess==0.70.16
mutagen==1.45.1
mypy==1.11.1
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
ninja==1.11.1.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
Pillow==9.0.1
pip-autoremove==0.10.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.0
protobuf==5.28.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.12.1
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
pygame==2.1.2
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
redis==5.2.0
referencing==0.35.1
regex==2024.9.11
rencode==1.0.6
repoze.lru==0.7
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.0
SecretStorage==3.3.1
Send2Trash==1.8.3
sentencepiece==0.2.0
service-identity==18.1.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
stemming==1.0.1
sympy==1.13.1
systemd-python==234
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.20.1
tomli==2.0.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
Twisted==22.1.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==9.1
Whoosh==2.7.4
widgetsnbextension==4.0.11
xdg==5
xxhash==3.5.0
yarl==1.18.0
yt-dlp==2022.4.8
zipp==1.0.0
zope.interface==5.4.0
